Refactor <Icon />
